LPD-46406 Moving virtual instances tests to a new module
