+++ modules/apps/portal/portal-virtual-instances-test/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Virtual Instances Test
+Bundle-SymbolicName: com.liferay.portal.virtual.instances.test
+Bundle-Version: 1.0.0
