+++ modules/apps/portal/portal-virtual-instances-test/build.gradle
@@ -1,0 +1,12 @@
+dependencies {
+	testIntegrationImplementation group: "com.liferay", name: "org.apache.felix.configadmin", version: "1.9.26.LIFERAY-PATCHED-1"
+	testIntegrationImplementation group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
+	testIntegrationImplementation project(":apps:portal-configuration:portal-configuration-test-util")
+	testIntegrationImplementation project(":apps:portal:portal-db-partition-test-util")
+	testIntegrationImplementation project(":apps:portal:portal-instance-lifecycle-api")
+	testIntegrationImplementation project(":apps:static:osgi:osgi-util")
+	testIntegrationImplementation project(":apps:static:portal-configuration:portal-configuration-metatype-api")
+	testIntegrationImplementation project(":apps:static:portal-configuration:portal-configuration-persistence-api")
+	testIntegrationImplementation project(":apps:static:portal-file-install:portal-file-install-api")
+	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
+}
