+++ modules/apps/portal/portal-virtual-instances-test/src/testIntegration/java/com/liferay/portal/virtual/instances/internal/operation/test/DBPartitionCopyVirtualInstanceOperationTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.db.partition.internal.operation.test;
+package com.liferay.portal.virtual.instances.internal.operation.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.string.StringBundler;
