+++ modules/apps/portal/portal-virtual-instances-test/src/testIntegration/java/com/liferay/portal/virtual/instances/internal/operation/test/DBPartitionExtractVirtualInstanceOperationTest.java
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.db.partition.internal.operation.test;
+package com.liferay.portal.virtual.instances.internal.operation.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.portal.kernel.instance.PortalInstancePool;
 import com.liferay.portal.test.log.LogCapture;
 import com.liferay.portal.test.log.LoggerTestUtil;
 
@@ -17,31 +16,26 @@ import org.junit.runner.RunWith;
  * @author Mariano Álvaro Sáiz
  */
 @RunWith(Arquillian.class)
-public class DBPartitionInsertVirtualInstanceOperationTest
+public class DBPartitionExtractVirtualInstanceOperationTest
 	extends BaseVirtualInstanceOperationTestCase {
 
 	@Override
 	public String getComponentName() {
-		return "InsertVirtualInstanceOperation";
+		return "ExtractVirtualInstanceOperation";
 	}
 
 	@Test
 	public void testDeployConfiguration() throws Exception {
 		try (LogCapture logCapture = LoggerTestUtil.configureLog4JLogger(
 				"com.liferay.portal.virtual.instances.internal.operation." +
-					"InsertVirtualInstanceOperation",
+					"ExtractVirtualInstanceOperation",
 				LoggerTestUtil.ERROR)) {
 
-			deployConfiguration(
-				_PID,
-				"newWebId=\"testNewWebId\"\ninsertCompanyId=L\"" +
-					PortalInstancePool.getDefaultCompanyId() + "\"\n");
+			deployConfiguration(_PID, "extractCompanyId=L\"0\"\n");
 
 			assertLog(
 				logCapture,
-				"Virtual instance with company ID " +
-					PortalInstancePool.getDefaultCompanyId() +
-						" already exists");
+				"Virtual instance with company ID 0 does not exist");
 		}
 
 		assertConfigurationIsDeletedAfterDeploy(_PID);
@@ -49,6 +43,6 @@ public class DBPartitionInsertVirtualInstanceOperationTest
 
 	private static final String _PID =
 		"com.liferay.portal.virtual.instances.internal.configuration." +
-			"InsertVirtualInstanceConfiguration";
+			"ExtractVirtualInstanceConfiguration";
 
 }
