+++ modules/apps/portal/portal-virtual-instances-test/test.properties
@@ -1,0 +1,5 @@
+##
+## Testray
+##
+
+    testray.main.component.name=Virtual Instances
